requirements.txt
python-dotenv
langchain-openai
langchain-core
pydantic
fastapi
langsmith